Move session list to a separate component
